apps/meta-agent/requirements.txt
llama-index==0.10.0
langroid==0.1.0
openai==1.3.6
pgvector==0.2.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
flask==2.3.2
requests==2.31.0
pydantic==2.0.3
pytesseract==0.3.10
pdf2image==1.16.3
PyJWT==2.8.0
redis==4.6.0
cryptography==41.0.0